lyss-provider-service/requirements.txt
# 核心框架
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic[email]==2.5.0
pydantic-settings==2.1.0

# 数据库
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.0

# Redis
redis==5.0.1
hiredis==2.3.2

# HTTP客户端
httpx==0.25.2
aiohttp==3.9.1

# 认证和安全
pyjwt==2.8.0
cryptography==41.0.8
passlib[bcrypt]==1.7.4

# AI供应商SDK
openai==1.3.7
anthropic==0.7.7
google-generativeai==0.3.2

# 配置和环境
python-dotenv==1.0.0
pyyaml==6.0.1

# 工具库
python-multipart==0.0.6
python-json-logger==2.0.7
tenacity==8.2.3
backoff==2.2.1

# 监控和指标
prometheus-client==0.19.0
structlog==23.2.0

# 开发工具
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# HTTP请求
requests==2.31.0